Adds a second shift vector to the fully_connected_16() function.
The second shift vector is built into the BSS tensor after the scale vector. The fully_connected_16() function now uses a VLMACCR to apply the scale instead of a VLMUL, and then the second shift vector is used with VLSAT to bring the result back down to 16 bits.

This change enables the fully_connected_16() function to scale accumulator results up by more than a factor of 2.
The required change to the assembly code added 4 additional instructions for each output row, column and channel group.