Fix load from flash to keep ordering of ops
Constants are added from the top of MLIR. This normally doesn't
matter, however we are replacing constants with ops to load them
from flash. Generalizing the op to replace all constants had the
unfortunate side effect of moving all load from flash ops to the
beginning of the graph. The generalization is removed and we are
specifically targeting xc_conv2d and tfl_conv2d.